requirements.txt
PyMuPDF
python-docx
bs4
pydub
moviepy
lm-format-enforcer
django
flask
googletrans
cython
nltk
torch
transformers
bitsandbytes
accelerate
nemo_toolkit[all]
git+https://github.com/huggingface/parler-tts.git
git+https://github.com/NVIDIA/NeMo.git@main#egg=nemo_toolkit[asr]
